+++ src/main/java/com/be3c/sysmetic/domain/member/dto/NoticeDetailAdminShowResponseDto.java
@@ -1,4 +1,4 @@
-package com.be3c.sysmetic.domain.member.dto.notice;
+package com.be3c.sysmetic.domain.member.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
@@ -10,8 +10,8 @@ import java.time.LocalDateTime;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-@Schema(description = "공지사항 상세 조회 응답 DTO")
-public class NoticeDetailShowResponseDto {
+@Schema(description = "관리자 공지사항 상세 조회 응답 DTO")
+public class NoticeDetailAdminShowResponseDto {
 
     @Schema(description = "공지사항 ID", example = "123")
     private Long noticeId;
@@ -36,4 +36,7 @@ public class NoticeDetailShowResponseDto {
 
     @Schema(description = "첨부 파일 여부 (0: 없음, 1: 있음)", example = "1")
     private Integer isAttatchment;
+
+    @Schema(description = "공개 여부 (0: 비공개, 1: 공개)", example = "1")
+    private Integer isOpen;
 }
